Initial commit of seat booking system
This commit includes the initial version of the Apache Airlines Seat Booking Application. The application provides the following functionalities:
- Check availability of seat
- Book a seat
- Free a seat
- Show booking state
- Exit program

The seating plan of the Burak757 includes rows labeled from 1 to 80 and columns labeled A to F, with specific seats designated as aisles ("X") and storage areas ("S"). Bookings are indicated by storing a booking reference in place of the letter "R".