Refactored code to include booking references and customer details.
Developed a function that will produce a random booking reference. The booking reference has exactly eight alphanumeric characters. When a new reference is produced the system makes sure the reference is unique (it has not been produced before).

Refactored the functionalities from part A in such a way that when a booking is made the reference is stored in the data structure and customer data is stored on in a data structure. Likewise, when a seat is freed the letter “F” is stored and any booking details from the database is removed.